requirements.txt
streamlit>=1.44.0
azure-storage-blob>=12.25.0
Flask>=3.1.0
python-dotenv>=1.0.1
